Testando com mais operações possíveis

+++ 2Período/Algoritmos2/exercicios/Trabalhos/Final/Questões 1 - 2/classes/multiplicaCast.java
@@ -9,7 +9,7 @@ public class multiplicaCast {
     }
 
     public int multiplica(int n1, int n2) {
-        return (int) m.multiplica(n1, n2);
+        return (int) m.multiplica(n1, (double) n2);
     }
 
     public int multiplica(float n1, float n2) {
